Handle streaming media of unknown length
The progress bar will now use the duration provided by the current
song's metadata instead of the duration of the `HTMLAudioElement`. This
allows for properly handling streaming media where the total size is
unknown, like when transcoding to a format without a consistent bitrate
(ie MP3 V0).

Since knowing the final size of the media is no longer required, the
`estimateContentLength` parameter was removed from the stream request.
Servers will still provide a `Content-Length` header for media where the
size is known, but will now refuse to guess if they're not sure.

Since it's possible that the duration in the song's metadata could be
missing, the duration of the `HTMLAudioElement` is still used as a
fallback. As more audio data is loaded, the duration of the element in
the state is updated.