NOMERGE: Add a "progress" bar for debugging
Note that this just for debugging. A progress bar of how much of a song
has been loaded would be a useful feature, but that's not what this
displays.

It shows the current duration of the loaded audio out of the duration of
the audio as reported by the song metadata.